# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/posts": {
      "post": {
        "operationId": "PostController_createPost",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreatePostDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ResponseDto" },
                    {
                      "properties": {
                        "message": { "example": "Created" },
                        "statusCode": { "type": "number", "example": 201 }
                      }
                    }
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/CreatePostResponseDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "message": { "example": "Not Found" },
                        "statusCode": { "type": "number", "example": 404 },
                        "timestamp": {
                          "type": "string",
                          "example": "2024-04-18T08:48:35.101Z"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": ["posts"],
        "security": [{ "accessToken": [] }]
      },
      "get": {
        "operationId": "PostController_getPosts",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "description": "limit",
            "example": 10,
            "schema": { "type": "number" }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "page",
            "example": 0,
            "schema": { "type": "number" }
          },
          {
            "name": "search",
            "required": true,
            "in": "query",
            "description": "search",
            "example": "civitas",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ResponseDto" },
                    {
                      "properties": {
                        "message": { "example": "OK" },
                        "statusCode": { "type": "number", "example": 200 }
                      }
                    }
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/GetPostsResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["posts"],
        "security": [{ "accessToken": [] }]
      }
    },
    "/api/v1/posts/{id}": {
      "delete": {
        "operationId": "PostController_deletePost",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ResponseDto" },
                    {
                      "properties": {
                        "message": { "example": "OK" },
                        "statusCode": { "type": "number", "example": 200 }
                      }
                    }
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/GenericResponseDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "message": { "example": "Not Found" },
                        "statusCode": { "type": "number", "example": 404 },
                        "timestamp": {
                          "type": "string",
                          "example": "2024-04-18T08:48:35.102Z"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": ["posts"],
        "security": [{ "accessToken": [] }]
      },
      "put": {
        "operationId": "PostController_updatePost",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdatePostDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ResponseDto" },
                    {
                      "properties": {
                        "message": { "example": "OK" },
                        "statusCode": { "type": "number", "example": 200 }
                      }
                    }
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/UpdatePostResponseDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "message": { "example": "Not Found" },
                        "statusCode": { "type": "number", "example": 404 },
                        "timestamp": {
                          "type": "string",
                          "example": "2024-04-18T08:48:35.102Z"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": ["posts"],
        "security": [{ "accessToken": [] }]
      }
    },
    "/api/v1/users/profile": {
      "get": {
        "operationId": "UserController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ResponseDto" },
                    {
                      "properties": {
                        "message": { "example": "OK" },
                        "statusCode": { "type": "number", "example": 200 }
                      }
                    }
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/GetProfileResponseDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "message": { "example": "Not Found" },
                        "statusCode": { "type": "number", "example": 404 },
                        "timestamp": {
                          "type": "string",
                          "example": "2024-04-18T08:48:35.069Z"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": ["user.users"],
        "security": [{ "accessToken": [] }]
      }
    },
    "/api/v1/users": {
      "put": {
        "operationId": "UserController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserUpdateDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ResponseDto" },
                    {
                      "properties": {
                        "message": { "example": "OK" },
                        "statusCode": { "type": "number", "example": 200 }
                      }
                    }
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/UpdateProfileResponseDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "message": { "example": "Not Found" },
                        "statusCode": { "type": "number", "example": 404 },
                        "timestamp": {
                          "type": "string",
                          "example": "2024-04-18T08:48:35.070Z"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": ["user.users"],
        "security": [{ "accessToken": [] }]
      }
    },
    "/api/v1/admin/users/{id}": {
      "delete": {
        "operationId": "AdminUserController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ResponseDto" },
                    {
                      "properties": {
                        "message": { "example": "OK" },
                        "statusCode": { "type": "number", "example": 200 }
                      }
                    }
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/GenericResponseDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "message": { "example": "Not Found" },
                        "statusCode": { "type": "number", "example": 404 },
                        "timestamp": {
                          "type": "string",
                          "example": "2024-04-18T08:48:35.098Z"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": ["admin.users"],
        "security": [{ "accessToken": [] }]
      },
      "put": {
        "operationId": "AdminUserController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserUpdateDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ResponseDto" },
                    {
                      "properties": {
                        "message": { "example": "OK" },
                        "statusCode": { "type": "number", "example": 200 }
                      }
                    }
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/UpdateProfileResponseDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "message": { "example": "Not Found" },
                        "statusCode": { "type": "number", "example": 404 },
                        "timestamp": {
                          "type": "string",
                          "example": "2024-04-18T08:48:35.098Z"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": ["admin.users"],
        "security": [{ "accessToken": [] }]
      }
    },
...
  },
  "info": {
    "title": "nestjs-starter APIs Specification",
    "description": "Section for description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "accessToken": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "refreshToken": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": { "type": "number", "example": 201 },
          "message": { "type": "string", "nullable": false },
          "timestamp": {
            "type": "string",
            "example": "2024-04-18T08:48:35.068Z",
            "nullable": false
          },
          "data": { "type": "object", "nullable": false }
        },
        "required": ["statusCode", "message", "timestamp", "data"]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "avatar_id": { "type": "string" },
          "created_at": { "format": "date-time", "type": "string" },
          "deleted_at": { "format": "date-time", "type": "string" },
          "email": { "type": "string" },
          "first_name": { "type": "string" },
          "id": { "type": "string" },
          "is_deleted": { "type": "boolean" },
          "is_verified": { "type": "boolean" },
          "last_name": { "type": "string" },
          "phone": { "type": "string" },
          "role": { "type": "object" },
          "updated_at": { "format": "date-time", "type": "string" },
          "password": { "type": "string" }
        },
        "required": [
          "avatar_id",
          "created_at",
          "deleted_at",
          "email",
          "first_name",
          "id",
          "is_deleted",
          "is_verified",
          "last_name",
          "phone",
          "role",
          "updated_at",
          "password"
        ]
      },
      "FileDto": {
        "type": "object",
        "properties": {
          "created_at": { "format": "date-time", "type": "string" },
          "deleted_at": { "format": "date-time", "type": "string" },
          "file_name": { "type": "string" },
          "id": { "type": "string" },
          "is_deleted": { "type": "boolean" },
          "link": { "type": "string" },
          "post_id": { "type": "string" },
          "type": { "type": "string" },
          "updated_at": { "format": "date-time", "type": "string" }
        },
        "required": [
          "created_at",
          "deleted_at",
          "file_name",
          "id",
          "is_deleted",
          "link",
          "post_id",
          "type",
          "updated_at"
        ]
      },
      "CreatePostResponseDto": {
        "type": "object",
        "properties": {
          "author_id": { "type": "string" },
          "content": { "type": "string" },
          "created_at": { "format": "date-time", "type": "string" },
          "deleted_at": { "format": "date-time", "type": "string" },
          "id": { "type": "string" },
          "is_deleted": { "type": "boolean" },
          "title": { "type": "string" },
          "updated_at": { "format": "date-time", "type": "string" },
          "author": { "$ref": "#/components/schemas/UserDto" },
          "photos": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/FileDto" }
          }
        },
        "required": [
          "author_id",
          "content",
          "created_at",
          "deleted_at",
          "id",
          "is_deleted",
          "title",
          "updated_at",
          "author",
          "photos"
        ]
      },
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "post content",
            "example": "Velum defero conturbo curia sollicito auctus ventus illo temporibus.\nVilis amitto substantia vere basium amplexus quisquam amplexus demonstro."
          },
          "title": {
            "type": "string",
            "description": "post title",
            "example": "aranea coaegresco adsuesco"
          }
        },
        "required": ["content", "title"]
      },
      "GenericResponseDto": {
        "type": "object",
        "properties": {
          "status": { "type": "boolean", "example": true },
          "message": { "type": "string", "example": "record is deleted" }
        },
        "required": ["status", "message"]
      },
      "PostResponseDto": {
        "type": "object",
        "properties": {
          "author_id": { "type": "string" },
          "content": { "type": "string" },
          "created_at": { "format": "date-time", "type": "string" },
          "deleted_at": { "format": "date-time", "type": "string" },
          "id": { "type": "string" },
          "is_deleted": { "type": "boolean" },
          "title": { "type": "string" },
          "updated_at": { "format": "date-time", "type": "string" },
          "author": { "$ref": "#/components/schemas/UserDto" },
          "photos": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/FileDto" }
          }
        },
        "required": [
          "author_id",
          "content",
          "created_at",
          "deleted_at",
          "id",
          "is_deleted",
          "title",
          "updated_at",
          "author",
          "photos"
        ]
      },
      "GetPostsResponseDto": {
        "type": "object",
        "properties": {
          "count": { "type": "number", "example": 10, "nullable": false },
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PostResponseDto" }
          }
        },
        "required": ["count", "data"]
      },
      "UpdatePostResponseDto": {
        "type": "object",
        "properties": {
          "author_id": { "type": "string" },
          "content": { "type": "string" },
          "created_at": { "format": "date-time", "type": "string" },
          "deleted_at": { "format": "date-time", "type": "string" },
          "id": { "type": "string" },
          "is_deleted": { "type": "boolean" },
          "title": { "type": "string" },
          "updated_at": { "format": "date-time", "type": "string" },
          "author": { "$ref": "#/components/schemas/UserDto" },
          "photos": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/FileDto" }
          }
        },
        "required": [
          "author_id",
          "content",
          "created_at",
          "deleted_at",
          "id",
          "is_deleted",
          "title",
          "updated_at",
          "author",
          "photos"
        ]
      },
      "UpdatePostDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "post content",
            "example": "Velum defero conturbo curia sollicito auctus ventus illo temporibus.\nVilis amitto substantia vere basium amplexus quisquam amplexus demonstro."
          },
          "title": {
            "type": "string",
            "description": "post title",
            "example": "aranea coaegresco adsuesco"
          }
        },
        "required": ["content", "title"]
      },
      "GetProfileResponseDto": {
        "type": "object",
        "properties": {
          "avatar_id": { "type": "string" },
          "created_at": { "format": "date-time", "type": "string" },
          "deleted_at": { "format": "date-time", "type": "string" },
          "email": { "type": "string" },
          "first_name": { "type": "string" },
          "id": { "type": "string" },
          "is_deleted": { "type": "boolean" },
          "is_verified": { "type": "boolean" },
          "last_name": { "type": "string" },
          "phone": { "type": "string" },
          "role": { "type": "object" },
          "updated_at": { "format": "date-time", "type": "string" },
          "password": { "type": "string" }
        },
        "required": [
          "avatar_id",
          "created_at",
          "deleted_at",
          "email",
          "first_name",
          "id",
          "is_deleted",
          "is_verified",
          "last_name",
          "phone",
          "role",
          "updated_at",
          "password"
        ]
      },
      "UpdateProfileResponseDto": {
        "type": "object",
        "properties": {
          "avatar_id": { "type": "string" },
          "created_at": { "format": "date-time", "type": "string" },
          "deleted_at": { "format": "date-time", "type": "string" },
          "email": { "type": "string" },
          "first_name": { "type": "string" },
          "id": { "type": "string" },
          "is_deleted": { "type": "boolean" },
          "is_verified": { "type": "boolean" },
          "last_name": { "type": "string" },
          "phone": { "type": "string" },
          "role": { "type": "object" },
          "updated_at": { "format": "date-time", "type": "string" },
          "password": { "type": "string" }
        },
        "required": [
          "avatar_id",
          "created_at",
          "deleted_at",
          "email",
          "first_name",
          "id",
          "is_deleted",
          "is_verified",
          "last_name",
          "phone",
          "role",
          "updated_at",
          "password"
        ]
      },
      "UserUpdateDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "user email address",
            "example": "Reagan.Swift@yahoo.com"
          },
          "firstName": {
            "type": "string",
            "description": "user first name",
            "example": "Sedrick"
          },
          "lastName": {
            "type": "string",
            "description": "user last name",
            "example": "Bahringer"
          },
          "profile": {
            "type": "string",
            "description": "user profile picture",
            "example": "https://avatars.githubusercontent.com/u/7709463"
          }
        }
      },
...
    }
  }
}